\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary}
\alias{summary,flowTransResult-method}
\title{ Summarize a flowTransResult object}
\description{
 Print a summary of the contents of a \code{flowTransResult} object.
}
\section{Methods}{
\describe{

\item{object = "flowTransResult"}{ Prints a summary of the flowTransResult object }
}}
\keyword{methods}
